refactor(charts): move ChartDataPoint into a shared types module
Pulls the ChartDataPoint type out of BarChart.tsx into
components/charts/types.ts so consumers (LineChart, UsageTrendsChart,
FeatureNavTab utils, story fake-data) don't have to reach into a
sibling component for the shape. Barrel re-export updated to point at
the new module.
