+++ frontend/web/components/charts/index.ts
@@ -1,7 +1,7 @@
 export { default as BarChart } from './BarChart'
-export type { ChartDataPoint } from './BarChart'
 export { default as LineChart } from './LineChart'
 export { default as PieChart } from './PieChart'
 export type { PieSlice } from './PieChart'
 export { default as ChartTooltip } from './ChartTooltip'
 export { buildChartColorMap } from './buildChartColorMap'
+export type { ChartDataPoint } from './types'
