+++ frontend/web/components/charts/types.ts
@@ -1,0 +1,4 @@
+export type ChartDataPoint = {
+  day: string
+  [key: string]: string | number
+}
